requirements.txt
behave==1.2.6
enum34==1.1.6
Faker==0.9.2
nose==1.3.7
parse==1.9.0
parse-type==0.4.2
PyHamcrest==1.9.0
PyYAML==3.13
selenium==3.141.0
six==1.11.0
splinter==0.9.0
